+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Cache data for user avatar in memory.
+- Better error feedback add verified water reading form.
+
+### Fixed
+- Add verified water readings required field.
+
 ## [1.7.2] - 2021-04-27
 ### Fixed
 - MinderGas.nl synchronization checkbox toggle.